// web/src/api/generated/models/ImportantDateView.ts
export interface ImportantDateView {
    /**
     * 
     * @type {Date}
     * @memberof ImportantDateView
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof ImportantDateView
     */
    date: Date;
    /**
     * 
     * @type {string}
     * @memberof ImportantDateView
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ImportantDateView
     */
    label: string;
    /**
     * 
     * @type {string}
     * @memberof ImportantDateView
     */
    relatedPersonId: string | null;
    /**
     * 
     * @type {DateRepeat}
     * @memberof ImportantDateView
     */
    repeats: DateRepeat;
    /**
     * 
     * @type {ImportantDateType}
     * @memberof ImportantDateView
     */
    type: ImportantDateType;
    /**
     * 
     * @type {Date}
     * @memberof ImportantDateView
     */
    updatedAt: Date;
    /**
     * 
     * @type {number}
     * @memberof ImportantDateView
     */
    version: number;
    /**
     * 
     * @type {ContentVisibility}
     * @memberof ImportantDateView
     */
    visibility: ContentVisibility;
}